Preserve chat user ids
Store the logged-in Twitch user ID in IrcChat as quint64 so Helix block-list loading and block edits receive the full user ID instead of a truncated 16-bit value. Extend the Twitch API guard to enforce the ID width.

+++ src/model/ircchat.h
@@ -161,7 +161,7 @@ private slots:
     void getBlockedUserList();
 
 private:
-    quint16 user_id;
+    quint64 user_id;
 
     static const qint16 PORT;
     static const QString HOST;
